move remote env stuff to own file
